[nrf fromtree] drivers: clock_control: Renamed nrf2 to nrf_common.
Renamed nrf2 to nrf_common to keep the name cosistent with its use.


(cherry picked from commit 082669e89e4544f7122415a14307a1f92ca88385)
