+++ drivers/clock_control/Kconfig.nrf
@@ -211,7 +211,12 @@ config CLOCK_CONTROL_NRF_ACCURACY
 
 endif # CLOCK_CONTROL_NRF
 
-config CLOCK_CONTROL_NRF2_COMMON
+config CLOCK_CONTROL_NRF_AUXPLL
+	bool "nRF Auxiliary PLL driver"
+	default y
+	depends on DT_HAS_NORDIC_NRF_AUXPLL_ENABLED
+
+config CLOCK_CONTROL_NRF_COMMON
 	bool
 	depends on HAS_NORDIC_DRIVERS
 	select ONOFF
@@ -221,7 +226,7 @@ config CLOCK_CONTROL_NRF_HSFLL_GLOBAL
 	depends on DT_HAS_NORDIC_NRF_HSFLL_GLOBAL_ENABLED
 	select NRFS
 	select NRFS_GDFS_SERVICE_ENABLED
-	select CLOCK_CONTROL_NRF2_COMMON
+	select CLOCK_CONTROL_NRF_COMMON
 	default y
 
 if CLOCK_CONTROL_NRF_HSFLL_GLOBAL
@@ -257,19 +262,19 @@ config CLOCK_CONTROL_NRFS_AUDIOPLL
 	depends on DT_HAS_NORDIC_NRFS_AUDIOPLL_ENABLED
 	select NRFS
 	select NRFS_AUDIOPLL_SERVICE_ENABLED
-	select CLOCK_CONTROL_NRF2_COMMON
+	select CLOCK_CONTROL_NRF_COMMON
 	default y
 
 config CLOCK_CONTROL_NRF_FLL16M
 	bool "NRF FLL16M driver support"
 	depends on DT_HAS_NORDIC_NRF_FLL16M_ENABLED
-	select CLOCK_CONTROL_NRF2_COMMON
+	select CLOCK_CONTROL_NRF_COMMON
 	default y
 
 config CLOCK_CONTROL_NRF54H_HFXO
 	bool "NRF54H HFXO driver support"
 	depends on DT_HAS_NORDIC_NRF54H_HFXO_ENABLED
-	select CLOCK_CONTROL_NRF2_COMMON
+	select CLOCK_CONTROL_NRF_COMMON
 	default y
 
 config CLOCK_CONTROL_NRF_HSFLL_LOCAL
@@ -277,7 +282,7 @@ config CLOCK_CONTROL_NRF_HSFLL_LOCAL
 	depends on DT_HAS_NORDIC_NRF_HSFLL_LOCAL_ENABLED
 	select NRFS
 	select NRFS_DVFS_LOCAL_DOMAIN
-	select CLOCK_CONTROL_NRF2_COMMON
+	select CLOCK_CONTROL_NRF_COMMON
 	default y
 
 if CLOCK_CONTROL_NRF_HSFLL_LOCAL
@@ -297,7 +302,7 @@ config CLOCK_CONTROL_NRF_IRON_HSFLL_LOCAL
 	bool "NRF IronSide HSFLL LOCAL driver support"
 	depends on DT_HAS_NORDIC_NRF_IRON_HSFLL_LOCAL_ENABLED
 	select IRONSIDE_SE_DVFS
-	select CLOCK_CONTROL_NRF2_COMMON
+	select CLOCK_CONTROL_NRF_COMMON
 	default y
 
 if CLOCK_CONTROL_NRF_IRON_HSFLL_LOCAL
@@ -318,7 +323,7 @@ config CLOCK_CONTROL_NRF_LFCLK
 	depends on DT_HAS_NORDIC_NRF_LFCLK_ENABLED
 	select NRFS
 	select NRFS_CLOCK_SERVICE_ENABLED
-	select CLOCK_CONTROL_NRF2_COMMON
+	select CLOCK_CONTROL_NRF_COMMON
 	default y
 
 if CLOCK_CONTROL_NRF_LFCLK
@@ -328,9 +333,3 @@ config CLOCK_CONTROL_NRF_LFCLK_CLOCK_TIMEOUT_MS
 	default 1000
 
 endif # CLOCK_CONTROL_NRF_LFCLK
-
-config CLOCK_CONTROL_NRF_AUXPLL
-	bool "nRF Auxiliary PLL driver"
-	default y
-	depends on DT_HAS_NORDIC_NRF_AUXPLL_ENABLED
-	select CLOCK_CONTROL_NRF2_COMMON
