+++ drivers/clock_control/clock_control_nrf_hsfll_global.c
@@ -5,13 +5,13 @@
 
 #define DT_DRV_COMPAT nordic_nrf_hsfll_global
 
-#include "clock_control_nrf2_common.h"
+#include "clock_control_nrf_common.h"
 #include <zephyr/devicetree.h>
 #include <zephyr/drivers/clock_control/nrf_clock_control.h>
 #include <nrfs_gdfs.h>
 
 #include <zephyr/logging/log.h>
-LOG_MODULE_DECLARE(clock_control_nrf2, CONFIG_CLOCK_CONTROL_LOG_LEVEL);
+LOG_MODULE_REGISTER(clock_control_nrf_hsfll_global, CONFIG_CLOCK_CONTROL_LOG_LEVEL);
 
 #define GLOBAL_HSFLL_CLOCK_FREQUENCIES \
 	DT_INST_PROP(0, supported_clock_frequencies)
