+++ drivers/clock_control/clock_control_nrf_lfclk.c
@@ -5,7 +5,7 @@
 
 #define DT_DRV_COMPAT nordic_nrf_lfclk
 
-#include "clock_control_nrf2_common.h"
+#include "clock_control_nrf_common.h"
 #include <zephyr/devicetree.h>
 #include <zephyr/drivers/clock_control/nrf_clock_control.h>
 #include <hal/nrf_bicr.h>
