+++ drivers/clock_control/clock_control_nrfs_audiopll.c
@@ -5,7 +5,7 @@
 
 #define DT_DRV_COMPAT nordic_nrfs_audiopll
 
-#include "clock_control_nrf2_common.h"
+#include "clock_control_nrf_common.h"
 #include <zephyr/devicetree.h>
 #include <zephyr/dt-bindings/clock/nrfs-audiopll.h>
 #include <zephyr/drivers/clock_control/nrf_clock_control.h>
@@ -13,7 +13,7 @@
 #include <nrfs_backend_ipc_service.h>
 
 #include <zephyr/logging/log.h>
-LOG_MODULE_DECLARE(clock_control_nrf2, CONFIG_CLOCK_CONTROL_LOG_LEVEL);
+LOG_MODULE_REGISTER(clock_control_nrf_hsfll_global, CONFIG_CLOCK_CONTROL_LOG_LEVEL);
 
 #define SHIM_DEFAULT_PRESCALER AUDIOPLL_DIV_12
